requirements.txt
# pip install -r requirements.txt
colorama==0.4.4
requests==2.26.0
openai~=1.36.1
python-dotenv~=1.0.1
telebot~=0.0.5